refactor(pods): merge pod statuses
