+++ pkg/controllers/launch_manager/launchmanager_controller.go
@@ -110,10 +110,10 @@ func (r *LaunchManagerReconciler) reconcileCheckStatus(ctx context.Context, inst
 		switch resources.HasLaunchInThisInstance(*instance, *robot) {
 		case true:
 
-			switch instance.Status.LaunchPodStatus.Created {
+			switch instance.Status.LaunchPodStatus.Status.Created {
 			case true:
 
-				switch instance.Status.LaunchPodStatus.Phase {
+				switch instance.Status.LaunchPodStatus.Status.Phase {
 				case corev1.PodRunning:
 
 					instance.Status.Phase = robotv1alpha1.LaunchManagerPhaseReady
@@ -127,7 +127,7 @@ func (r *LaunchManagerReconciler) reconcileCheckStatus(ctx context.Context, inst
 				if err != nil {
 					return err
 				}
-				instance.Status.LaunchPodStatus.Created = true
+				instance.Status.LaunchPodStatus.Status.Created = true
 				instance.Status.Phase = robotv1alpha1.LaunchManagerPhaseLaunching
 
 			}
